PlayerInit: Move to `src/player/PlayerInit.js`
