+++ src/player/PlayerInit.js
@@ -13,7 +13,7 @@
  * @link Github - https://github.com/andriycraft/GreenFrogMCBE
  * @link Discord - https://discord.gg/UFqrnAbqjP
  */
-const Logger = require("./Logger");
+const Logger = require("../server/Logger");
 
 const Gamemode = require("../api/player/Gamemode");
